Clean up test classes
